repro for https://github.com/ansible/ansible/issues/22985
the `roles:` keyword cannot take a list that is not specified inline.  that is, it cannot be a variable.

```
- hosts: localhost
  gather_facts: false

  vars:
    whatroles:
      - role1
      - role2
      - role3

  roles: '{{whatroles}}'
```

execute the give list of roles

```
ERROR! A malformed role declaration was encountered.

The error appears to have been in '/home/scott/src/setup/test/test.yml': line 4, column 3, but may
be elsewhere in the file depending on the exact syntax problem.

The offending line appears to be:

- hosts: all
  ^ here

```